Weitere Bindings, sowie die Klassen-/Gruppen-/Managerklassen generalisiert
git-svn-id: https://aquitaine.dnsalias.net/repo/svn/shitstorm/trunk@220 89e8501b-ea7b-4f45-b1b2-b812960b401a

+++ Vector/GameLogic/EngineBase.h
@@ -255,7 +255,7 @@ namespace pv {
 		inline void setPhysicsFixedTimestep(irr::f32 step) { physicsFixedTimestep = step; }
 
 		//! Größe des fixen Timesteps in Bullet
-		inline irr::u32 getPhysicsUpdateFactor() const { return physicsUpdateFactor; }
+		inline irr::f32 getPhysicsUpdateFactor() const { return physicsUpdateFactor; }
 
 		//! Größe des fixen Timesteps in Bullet
 		inline void setPhysicsUpdateFactor(irr::f32 factor) { physicsUpdateFactor = factor; }
